Pick up and drop cards

+++ controller.gd
@@ -7,7 +7,8 @@ extends Node
 
 func on_interact_empty():
     print("clicked empty")
-    holding.drop()
+    if not holding.is_empty():
+        holding.drop()
 
 
 func on_interact(clicked: Node2D):
@@ -18,7 +19,9 @@ func on_interact(clicked: Node2D):
             c.set_face_up(true)
         stock.add_cards(drawn)
     if clicked is Card:
-        pass
+        if holding.is_empty():
+            var picked_up: Array[Card] = clicked.part_of.begin_move(clicked)
+            holding.pick_up(picked_up)
 
 
 func on_auto_move(clicked: Node2D):
